requirements.txt
streamlit
geopandas
leafmap
folium
python-docx
pyarrow
fiona